made a few improvements
add $ to the end of the search query to hide screenshots for "lite mode". make sure to have ^^^$ as the query if you want to view all posts! IF NOT, DOOM COMES. I kid, I kid. It just searches for "$".

also, I have replaced search_query= with q= since that makes a million times more sense

yeah that's about it.